Fix progress loading indicator in RefreshScreen

+++ appcompose/src/main/java/com/readrops/app/compose/util/components/RefreshScreen.kt
@@ -13,7 +13,7 @@ fun RefreshScreen(
 ) {
     CenteredColumn {
         LinearProgressIndicator(
-            progress = { (feedCount / feedMax).toFloat() }
+            progress = { feedCount.toFloat() / feedMax.toFloat() }
         )
 
         VeryShortSpacer()
